requirements.txt
lask==3.0.0
flask-cors==4.0.0
playwright==1.40.0
langgraph==0.2.27
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
fpdf2==2.8.5
